apps/iTung/backend_backup/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
anthropic
python-dotenv
bcrypt
python-jose
pydantic[email]
python-multipart
boto3
matplotlib
numpy
pillow
google-auth
cryptography
requests
user-agents
slowapi
sympy
